--- hooks/getProfilePicture.js
@@ -1,9 +1,0 @@
-export function getProfilePicture(contact) {
-  const path = "/profile-pictures/";
-
-  const { profilePicture } = contact;
-
-  if (profilePicture) {
-    return path + profilePicture;
-  }
-}
